pyproject.toml
[project]
name = "tombola"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "streamlit>=1.37.0",
    "google-api-python-client>=2.144.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
